Implemented GPSMapBranchV
Render at a per-brach basis instead
